Pass through all args/kwargs in Grain.__init__
In verion 2023.7.11 pymatgen added a new keyword argument to the
Structure constructor.  Grain doesn't know about it, so Grain.from_sites would break with this or later version of pymatgen.

I decided to pass everything through as *args/**kwargs because would
require a version bump of the pymatgen dependency as it wouldn't be
backwards compatible.